Temp File Fix and PDF Compile Changes - Fixed a bug where temp files were not properly deleted - Fixed holiday time formatting when exporting PDF - Fixed issues regarding duplicate, edit and remove buttons for selection by using implemented methods for getting the selected entry instead - Implemented Signature Option

+++ UITimesheetGenerator/src/main/java/net/justonedev/kit/export/TempFiles.java
@@ -9,7 +9,7 @@ import net.justonedev.kit.json.JSONHandler;
 
 import java.io.File;
 
-class TempFiles {
+class TempFiles implements AutoCloseable {
 
     private final File globalFile;
     private final File monthFile;
@@ -30,9 +30,11 @@ class TempFiles {
         return monthFile;
     }
 
+    @Override
     public void close() {
         if (isTempMonthFile) {
-            monthFile.deleteOnExit();
+            if (!monthFile.delete())
+                monthFile.deleteOnExit();
         }
     }
